<Intermediate commit> Test project changes - used TDD approach to configure Database project
