+++ VirtualLibraryTests/Tests/BookTest.cs
