+++ VirtualLibraryTests/VirtualLibraryTests.csproj
@@ -42,6 +42,7 @@
     <Reference Include="Castle.Core, Version=4.0.0.0, Culture=neutral, PublicKeyToken=407dd0808d44fbdc, processorArchitecture=MSIL">
       <HintPath>..\packages\Castle.Core.4.3.1\lib\net45\Castle.Core.dll</HintPath>
     </Reference>
+    <Reference Include="EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
     <Reference Include="Microsoft.VisualStudio.TestPlatform.TestFramework, Version=14.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
       <HintPath>..\packages\MSTest.TestFramework.1.3.2\lib\net45\Microsoft.VisualStudio.TestPlatform.TestFramework.dll</HintPath>
     </Reference>
@@ -53,6 +54,8 @@
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Configuration" />
+    <Reference Include="System.Data" />
+    <Reference Include="System.Data.Entity" />
     <Reference Include="System.Threading.Tasks.Extensions, Version=4.1.0.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL">
       <HintPath>..\packages\System.Threading.Tasks.Extensions.4.3.0\lib\portable-net45+win8+wp8+wpa81\System.Threading.Tasks.Extensions.dll</HintPath>
     </Reference>
@@ -69,9 +72,10 @@
     <Otherwise />
   </Choose>
   <ItemGroup>
-    <Compile Include="Helpers\BookTest.cs" />
-    <Compile Include="Helpers\InputValidatorTests.cs" />
-    <Compile Include="Helpers\UserPresenterTest.cs" />
+    <Compile Include="Tests\BookTest.cs" />
+    <Compile Include="Tests\InputValidatorTests.cs" />
+    <Compile Include="Tests\LibraryDbContextTest.cs" />
+    <Compile Include="Tests\UserPresenterTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -79,6 +83,10 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\Database\Database.csproj">
+      <Project>{E3089982-E97E-445D-A593-093209F31BFE}</Project>
+      <Name>Database</Name>
+    </ProjectReference>
     <ProjectReference Include="..\VirtualLibrary\VirtualLibrary.csproj">
       <Project>{6F784BA1-E19D-48C1-87A5-0A4C8ECB3548}</Project>
       <Name>VirtualLibrary</Name>
